dfeed.web: Allow reposting drafts approved by moderators
Summitted drafts are saved to a map for the purposes of deduplication:
if a user submits a draft which is identical to a draft that has been
submitted before, then we simply redirect to the previous post.

The effect of this persists even if a draft got as far as having a
PostProcess created for it, but is then flagged for moderation.  This
can happen if it fails a spam check and there is no CAPTCHA configured
for this group.  Effectively, the result of submitted a post is
memoized.

However, this deduplication does not have the intended effect when a
moderator approves a draft that has been prevented from being posted.
The memoization does not account for the new variable of the
moderator's approving override.

To remedy this, delete the draft from the deduplication AA before
re-submitting it, which allows PostProcess to proceed as with a new
post.